requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and NLP
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
openai-whisper==20231117
nltk==3.8.1

# Audio processing
librosa==0.10.1
soundfile==1.0.7
pyaudio==0.2.11

# Text-to-Speech
gTTS==2.4.0
pydub==0.25.1

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# HTTP and API clients
httpx==0.25.2
requests==2.31.0

# Database and caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Configuration and utilities
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# GUI Framework (optional - install with: pip install -r gui_requirements.txt)
# streamlit>=1.28.0
# plotly>=5.15.0
# pandas>=2.0.0